Improve IPC broadcast performance
The broadcast was initially sent in serialized (ordered) mode, which
could cause delays when Android is very busy (such as at early boot).
The broadcast was changed to use normal broadcast mode.

Additionally, FLAG_RECEIVER_FOREGROUND was added to the intent flags,
to make sure the receiver gets a foreground priority boost.