Daemonize: additional constructor options: surviveFrameworkRestart, exitListener
surviveFrameworkRestart (recommended: false) controls whether the daemon
automatically terminates if the Android framework dies. This is commonly
done during testing and by some root apps by calling stop and start in a
root shell. Setting this to true doesn't guarantee your daemon will
continue to function properly through a framework restart, that depends
on the rest of your code - it just won't automatically terminate.

exitListener is a callback for when the daemon is terminated through
the RootDaemon.exit() call. It IS NOT fired when the daemon is force
killed, when you call System.exit(), or an unhandled exception occurs.
It IS fired when you manually call RootDaemon.exit(), when
the daemon is being replaced by a newer version, and when the daemon
terminates due to framework death.